+++ python/ray/rllib/models/catalog.py
@@ -266,7 +266,7 @@ class _RLlibPreprocessorWrapper(gym.ObservationWrapper):
         self.preprocessor = preprocessor
 
         from gym.spaces.box import Box
-        self.observation_space = Box(-1.0, 1.0, preprocessor.shape)
+        self.observation_space = Box(-1.0, 1.0, preprocessor.shape, dtype=np.float32)
 
     def observation(self, observation):
         return self.preprocessor.transform(observation)
